% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluNorm.R
\name{fluNorm}
\alias{fluNorm}
\title{Title}
\usage{
fluNorm(pr_data, neg_well, flu_name)
}
\arguments{
\item{pr_data}{a long data.frame containing you plate reader data with OD normalised}

\item{neg_well}{the well coordinates of a non-fluorescent control}

\item{flu_name}{the column name of the fluorescence chanel to normalise}
}
\value{

}
\description{
Title
}
