% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odNorm.R
\name{odNorm}
\alias{odNorm}
\title{Title}
\usage{
odNorm(pr_data, blank_well, OD_name)
}
\arguments{
\item{pr_data}{a long data.frame containing you plate reader data}

\item{blank_well}{the well coordinates of a media blank}

\item{OD_name}{the column name for the optical density data}
}
\value{
an updated data.frame with an additional column "normalised_OD"
}
\description{
Title
}
