% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prNorm.R
\name{prNorm}
\alias{prNorm}
\title{Title}
\usage{
prNorm(pr_data, blank_well = "A1", neg_well = "A2", OD_name = "OD",
  flu_names = c("GFP"))
}
\arguments{
\item{pr_data}{a long data.frame containing you plate reader data}

\item{blank_well}{the well coordinates of a media blank}

\item{neg_well}{the well coordinates of a non-fluorescent control}

\item{OD_name}{the column name for the optical density data}

\item{flu_names}{the column names for the fluorscence data}
}
\description{
Title
}
