% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toMEF.R
\name{toMEF}
\alias{toMEF}
\title{Get Conversion Factors
Make sure the ggplot2 library is installed.}
\usage{
toMEF(pr_data, lid, GFP_gain, OD_name, conversion_factors_csv)
}
\arguments{
\item{lid}{if to_MEFL=T, the sealing mechanism used on the microtitre-plate. See the conversion_factors_csv for available options}

\item{GFP_gain}{if to_MEFL=T, the gain value at which GFP was recorded}

\item{OD_name}{the column name for the optical density data}

\item{conversion_factors_csv}{if to_MEFL=T, path of the csv file containing conversion factors from plate reader calibration}
}
\value{

}
\description{
Get Conversion Factors
Make sure the ggplot2 library is installed.
}
